requirements.txt
discord.py[voice]==2.6.4
python-dotenv==1.0.0
anthropic==0.34.2
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-storage==2.10.0
aiofiles==23.2.1
schedule==1.2.0
aiohttp>=3.13.2
beautifulsoup4==4.12.2
httpx==0.25.2
requests==2.31.0
matplotlib==3.8.2
flask==3.0.0
flask-cors==4.0.0
psutil==5.9.8
# Voice & TTS
PyNaCl==1.5.0
gTTS==2.5.1
pydub==0.25.1